requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Django==4.2.5
djangorestframework==3.14.0
kombu==5.3.2
numpy==1.26.0
packaging==23.1
pandas==2.1.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.3.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
